fix precompile for first time

+++ lib/capobvious/recipes/assets.rb
@@ -13,8 +13,8 @@ Capistrano::Configuration.instance(:must_exist).load do
     end
    desc "Assets precompile"
    task :precompile, :roles => :web, :except => { :no_release => true } do
-     if exists?(:turbo_sprockets_rails3) && fetch(:turbo_sprockets_rails3) && previous_release != latest_release
-       logger.info "Copying assets folder of previous release for turbo-sprockets-rails3"
+     if exists?(:turbo_sprockets_rails3) && fetch(:turbo_sprockets_rails3) && fetch(:previous_release) && previous_release != latest_release
+       logger.info "Copying assets folder of previous release for turbo-sprockets-rails3 gem"
        run("cp -r #{previous_release}/public/assets #{latest_release}/public/")
      end
      t1 = Time.now
